# Remove `torch` dependency, Faster numpy Feature extraction