--- a/requirements.txt
+++ b/requirements.txt
@@ -2,6 +2,5 @@ ctranslate2>=4.0,<5
 huggingface_hub>=0.13
 tokenizers>=0.13,<1
 onnxruntime>=1.14,<2 
-torch>=2.1.1 
 av>=11
 tqdm