@ozancaglayan 2024-10-31
Hi,

What's the value for `MKL_NUM_THREADS` and `OMP_NUM_THREADS` here for the CPU tests? Or are they not set?