@ozancaglayan 2024-10-31
I have a `librosa.stft` based implementation for this. I checked that your `stft` and `librosa.stft` is producing the same results and more or less performing at the same speed (6ms) Unfortunately, these implementations does not seem to be using multi-threaded computation whereas `torch.stft` efficiently uses all cores which make the difference: 1.6ms

Another main bottleneck here is the padding with 30 seconds. 

- I get `32ms` on a sample of 30s of audio with `n_mels=128` and `27ms` with `n_mels=80`
- With `padding=False`, `32ms` goes down to `16ms` obviously

But essentially the effect of padding the input is padding the output with some constant vector. So one could mitigate this doubling by just padding the stft/mel outputs cleverly.