@ozancaglayan 2024-10-31
This looks interesting as well, should be quite fast
https://github.com/IntelPython/mkl_fft