@MahmoudAshraf97 2024-10-31
> Hi,
> 
> * What's the value for `MKL_NUM_THREADS` and `OMP_NUM_THREADS` here for the CPU tests? Or are they not set?
> * Is this for 80 mels or 128 mels case?

I updated the figures along with the benchmarking script and limiting it to 4 cores for reproducability
I'm using 128 Mels

> I think doing this on GPU with torch is a bit overkill as feature extraction is seldom the bottleneck in the whisper pipeline. Simplicity, dependency management is more important here. And when things move towards numpy2, it'll be not that bad.

We're removing torch that's for sure, the choice here is between both cupy and numpy or one of them



> I have a `librosa.stft` based implementation for this. I checked that your `stft` and `librosa.stft` is producing the same results and more or less performing at the same speed (6ms for unpadded 30s of audio). Unfortunately, these implementations does not seem to be using multi-threaded computation whereas `torch.stft` efficiently uses all cores which make the difference: `~1ms`
> 

This implementation is ported from C++ pytorch implementation, it's very similar to the librosa implementation but has some extra functionality that I'm not sure we need here

> Another main bottleneck here is the padding with 30 seconds.
> 
> * I get `32ms` on a sample of 30s of audio with `n_mels=128` and `27ms` with `n_mels=80`
> * With `padding=False`, `32ms` goes down to `16ms` obviously
> 
> But essentially the effect of padding the input is padding the output with some constant vector. So one could mitigate this doubling by just padding the stft/mel outputs cleverly.


I'm working on removing the padding completely while still generating identical results but I'm leaving it for another PR, [Check here](https://github.com/huggingface/transformers/pull/34111#issuecomment-2441321749) it does reduce the FE time to half indeed

> We lose almost the same amount of time as STFT to `np.log10(np.clip(mels, a_min=1e-10, a_max=None))` as `np.log10` is horribly slow. I think they fixed this in numpy 2 [numpy/numpy#19478](https://github.com/numpy/numpy/pull/19478)

I have no idea how to calculate `log10` using another way that is faster unfortunately