@ozancaglayan 2024-10-31
> I'm working on removing the padding completely while still generating identical results but I'm leaving it for another PR, https://github.com/huggingface/transformers/pull/34111#issuecomment-2441321749 it does reduce the FE time to half indeed

Interesting. I'm always confused about one thing. Given that whisper is trained with 30 secs segments, the above PR and your thinking is not about running Whisper freely on arbitrarily sized audio files but rather removing the `always add 30 secs more of 0.0s` logic in FE right?

I thought that to do proper seeking and timestamping, Whisper requires that +30 secs of zeros all the time but isn't that the case then? i.e. would the output of Whisper be exactly the same if we just pad the inputs to 30 secs at most?