@MahmoudAshraf97 2024-10-31
I implemented the reduced padding in bc86503, everyone feel free to test it and report if there are any issues